+++ src/main/java/roomescape/controller/AdminPageController.java
@@ -1,0 +1,17 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminPageController {
+    @GetMapping("/admin")
+    public String admin() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String adminReservation() {
+        return "admin/reservation-legacy";
+    }
+}
