+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,45 +1,38 @@
 package roomescape.controller;
 
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.ReservationRepository;
 import roomescape.dto.AddReservationDto;
 import roomescape.model.Reservation;
 
-@Controller
-public class RoomescapeController {
+@RestController
+@RequestMapping("/reservations")
+public class ReservationController {
+    private final ReservationRepository reservationRepository;
 
-    @Autowired
-    ReservationRepository reservationRepository;
-
-    @GetMapping("/admin")
-    public String admin() {
-        return "admin/index";
+    public ReservationController(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
     }
 
-    @GetMapping("/admin/reservation")
-    public String adminReservation() {
-        return "admin/reservation-legacy";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping()
     public ResponseEntity<List<Reservation>> reservations() {
         return ResponseEntity.ok(reservationRepository.getAllReservations());
     }
 
-    @PostMapping("/reservations")
+    @PostMapping()
     public ResponseEntity<Reservation> addReservation(@RequestBody AddReservationDto addReservationDto) {
         return ResponseEntity.ok(reservationRepository.addReservation(addReservationDto));
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Integer> deleteReservation(@PathVariable long id) {
         return ResponseEntity.ok().body(reservationRepository.deleteReservation(id));
     }
